Shopify Service/requirements.txt
Flask-SQLAlchemy
Flask 
python-dotenv
requests
flask_swagger_ui
flask_restx